Suppression Inventory.js apres refactoring
